+++ packages/button/src/Component.tsx
@@ -12,6 +12,13 @@ import { useFocus } from '@alfalab/hooks';
 import { Loader } from '@alfalab/core-components-loader';
 
 import styles from './index.module.css';
+import defaultColors from './default.module.css';
+import invertedColors from './inverted.module.css';
+
+const colorStyles = {
+    default: defaultColors,
+    inverted: invertedColors,
+};
 
 export type ComponentProps = {
     /**
@@ -63,6 +70,11 @@ export type ComponentProps = {
      * Не переносить текст кнопки на новую строку
      */
     nowrap?: boolean;
+
+    /**
+     * Набор цветов для компонента
+     */
+    colors?: 'default' | 'inverted';
 };
 
 type AnchorButtonProps = ComponentProps & AnchorHTMLAttributes<HTMLAnchorElement>;
@@ -90,6 +102,7 @@ export const Button = React.forwardRef<HTMLAnchorElement | HTMLButtonElement, Bu
             href,
             loading = false,
             nowrap = false,
+            colors = 'default',
             ...restProps
         },
         ref,
@@ -109,12 +122,15 @@ export const Button = React.forwardRef<HTMLAnchorElement | HTMLButtonElement, Bu
                 styles.component,
                 styles[view],
                 styles[size],
+                colorStyles[colors].component,
+                colorStyles[colors][view],
                 {
                     [styles.focused]: focused,
                     [styles.block]: block,
                     [styles.iconOnly]: !children,
                     [styles.nowrap]: nowrap,
                     [styles.loading]: showLoader,
+                    [colorStyles[colors].loading]: showLoader,
                 },
                 className,
             ),