+++ packages/confirmation/src/component.tsx
@@ -80,7 +80,7 @@ export type ConfirmationProps = {
     /**
      * Заголовок экрана подписания
      */
-    signTitle?: string;
+    signTitle?: string | React.ReactNode;
 
     /**
      * Заголовок экрана блокирующей ошибки
@@ -177,7 +177,7 @@ export const Confirmation = forwardRef<HTMLDivElement, ConfirmationProps>(
             codeCheckingText = 'Проверка кода',
             codeSendingText = 'Отправляем код',
             buttonErrorText = 'Понятно',
-            buttonRetryText = 'Попробовать заново',
+            buttonRetryText = 'Вернуться назад',
             alignContent = 'left',
             noAttemptsLeftMessage,
             onInputFinished,
@@ -281,12 +281,7 @@ export const Confirmation = forwardRef<HTMLDivElement, ConfirmationProps>(
 
                         <span className={styles.errorText}>{errorText}</span>
 
-                        <Button
-                            size='s'
-                            view='secondary'
-                            onClick={handleErrorSmsRetryClick}
-                            block={true}
-                        >
+                        <Button size='xs' view='secondary' onClick={handleErrorSmsRetryClick}>
                             {buttonErrorText}
                         </Button>
                     </div>
@@ -327,9 +322,8 @@ export const Confirmation = forwardRef<HTMLDivElement, ConfirmationProps>(
 
                         <Button
                             className={styles.repeatButton}
-                            size='s'
+                            size='xs'
                             view='secondary'
-                            block={true}
                             onClick={handleSmsRetryFromHintClick}
                         >
                             {buttonRetryText}
@@ -356,6 +350,6 @@ Confirmation.defaultProps = {
     codeCheckingText: 'Проверка кода',
     codeSendingText: 'Отправляем код',
     buttonErrorText: 'Понятно',
-    buttonRetryText: 'Попробовать заново',
+    buttonRetryText: 'Вернуться назад',
     alignContent: 'left',
 };