+++ packages/confirmation/src/components/sign-confirmation/component.tsx
@@ -22,7 +22,7 @@ export type SignConfirmationProps = {
     code: string;
     errorText: string;
     error: boolean;
-    title: string;
+    title: string | React.ReactNode;
     codeCheckingText: string;
     codeSendingText: string;
     hasSmsCountdown: boolean;
@@ -98,7 +98,7 @@ export const SignConfirmation: FC<SignConfirmationProps> = ({
 
     return (
         <div className={cn(styles.component, styles[alignContent])}>
-            <span className={styles.header}>{title}</span>
+            {typeof title === 'string' ? <span className={styles.header}>{title}</span> : title}
 
             <div className={styles.inputContainer}>
                 <CodeInput
@@ -127,14 +127,14 @@ export const SignConfirmation: FC<SignConfirmationProps> = ({
             )}
 
             {hasSmsCountdown && (
-                <div className={cn('countdown', { [styles.hidden]: processing })}>
+                <div className={cn('countdown', styles.countdown, { [styles.hidden]: processing })}>
                     <Countdown
                         duration={countdownDuration}
                         phone={phone}
                         hasPhoneMask={hasPhoneMask}
-                        className={styles.countdown}
                         alignContent={alignContent}
                         noAttemptsLeftMessage={noAttemptsLeftMessage}
+                        hasError={Boolean(displayedError)}
                         onRepeatSms={onSmsRetryClick}
                         onCountdownFinished={onCountdownFinished}
                     />
@@ -147,6 +147,7 @@ export const SignConfirmation: FC<SignConfirmationProps> = ({
                         onClick={onSmsHintLinkClick}
                         className={styles.smsComeLink}
                         view='secondary'
+                        pseudo={true}
                     >
                         Не приходит сообщение?
                     </Link>