+++ packages/modal/src/components/content/Component.desktop.tsx
@@ -9,8 +9,22 @@ export type ContentDesktopProps = ContentProps & {
      * Размер
      */
     size?: 's' | 'm' | 'l';
+
+    /**
+     * Флаг, что модальное окно открыто на весь экран
+     */
+    fullscreen?: boolean;
 };
 
-export const ContentDesktop: FC<ContentDesktopProps> = ({ size, className, ...restProps }) => (
-    <Content className={cn(className, size && styles[size])} styles={styles} {...restProps} />
+export const ContentDesktop: FC<ContentDesktopProps> = ({
+    size,
+    className,
+    fullscreen,
+    ...restProps
+}) => (
+    <Content
+        className={cn(className, size && styles[size], fullscreen && styles.fullscreen)}
+        styles={styles}
+        {...restProps}
+    />
 );