+++ packages/dropzone/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.1](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-dropzone@1.1.0...@alfalab/core-components-dropzone@1.1.1) (2021-06-04)
+
+
+### Bug Fixes
+
+* **dropzone:** add pointer-events none to overlay ([#677](https://github.com/alfa-laboratory/core-components/issues/677)) ([5f9077a](https://github.com/alfa-laboratory/core-components/commit/5f9077a7c6c88272c7e9e67ec243488249cc5ded))
+
+
+
+
+
 # 1.1.0 (2021-04-26)