+++ packages/intl-phone-input/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.1.5](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-intl-phone-input@4.1.4...@alfalab/core-components-intl-phone-input@4.1.5) (2021-06-04)
+
+**Note:** Version bump only for package @alfalab/core-components-intl-phone-input
+
+
+
+
+
+## [4.1.4](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-intl-phone-input@4.1.3...@alfalab/core-components-intl-phone-input@4.1.4) (2021-05-31)
+
+**Note:** Version bump only for package @alfalab/core-components-intl-phone-input
+
+
+
+
+
+## [4.1.3](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-intl-phone-input@4.1.2...@alfalab/core-components-intl-phone-input@4.1.3) (2021-05-25)
+
+**Note:** Version bump only for package @alfalab/core-components-intl-phone-input
+
+
+
+
+
+## [4.1.2](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-intl-phone-input@4.1.1...@alfalab/core-components-intl-phone-input@4.1.2) (2021-05-25)
+
+**Note:** Version bump only for package @alfalab/core-components-intl-phone-input
+
+
+
+
+
 ## [4.1.1](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-intl-phone-input@4.1.0...@alfalab/core-components-intl-phone-input@4.1.1) (2021-05-18)
 
 **Note:** Version bump only for package @alfalab/core-components-intl-phone-input