+++ packages/typography/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.5.0](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-typography@1.4.7...@alfalab/core-components-typography@1.5.0) (2021-06-04)
+
+
+### Features
+
+* **typography:** add rest props ([#682](https://github.com/alfa-laboratory/core-components/issues/682)) ([51e1cf8](https://github.com/alfa-laboratory/core-components/commit/51e1cf89028c788bfcdbc21ea16acbbe190b5db9))
+
+
+
+
+
 ## [1.4.7](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-typography@1.4.6...@alfalab/core-components-typography@1.4.7) (2021-04-26)
 
 **Note:** Version bump only for package @alfalab/core-components-typography